rename modifiers.h to attribute_node_replacer.h
