+++ searchlib/src/vespa/searchlib/aggregation/CMakeLists.txt
@@ -3,6 +3,7 @@ vespa_add_library(searchlib_aggregation OBJECT
     SOURCES
     aggregation.cpp
     argmin_aggregation_result.cpp
+    attribute_node_replacer.cpp
     averageaggregationresult.cpp
     countaggregationresult.cpp
     expressioncountaggregationresult.cpp
@@ -15,7 +16,6 @@ vespa_add_library(searchlib_aggregation OBJECT
     hitsaggregationresult.cpp
     maxaggregationresult.cpp
     minaggregationresult.cpp
-    modifiers.cpp
     quantile_aggregation_result.cpp
     rawrank.cpp
     standarddeviationaggregationresult.cpp
