+++ searchlib/src/vespa/searchlib/aggregation/attribute_node_replacer.cpp
@@ -1,6 +1,6 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "modifiers.h"
+#include "attribute_node_replacer.h"
 
 #include "grouping.h"
